+++ tests/test_pandas.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from numpy.random import randn
-from nose.tools import assert_equal, assert_raises, with_setup
+from nose.tools import assert_equal, assert_raises, with_setup, assert_true
 
 import craftai.pandas
 from .data import pandas_valid_data
@@ -137,10 +137,13 @@ def test_get_operations_list_df_complex_agent():
 @with_setup(setup_complex_agent_with_data, teardown)
 def test_decide_from_contexts_df():
   tree = CLIENT.get_decision_tree(AGENT_ID, COMPLEX_AGENT_DATA.last_valid_index().value // 10 ** 9)
-  df = CLIENT.decide_from_contexts_df(tree, COMPLEX_AGENT_DATA)
+  test_df = COMPLEX_AGENT_DATA
+  test_df_copy = test_df.copy(deep=True)
+  df = CLIENT.decide_from_contexts_df(tree, test_df)
 
   assert_equal(len(df), 10)
   assert_equal(len(df.dtypes), 3)
+  assert_true(test_df.equals(test_df_copy))
   assert_equal(df.first_valid_index(), pd.Timestamp("2013-01-01 00:00:00", tz="Europe/Paris"))
   assert_equal(df.last_valid_index(), pd.Timestamp("2013-01-10 00:00:00", tz="Europe/Paris"))
 
@@ -151,6 +154,17 @@ def test_decide_from_contexts_df():
   })
 
   assert_equal(output["output"]["b"]["predicted_value"], "Pierre")
+
+@with_setup(setup_complex_agent_with_data, teardown)
+def test_decide_from_contexts_df_zero_rows():
+  tree = CLIENT.get_decision_tree(AGENT_ID, COMPLEX_AGENT_DATA.last_valid_index().value // 10 ** 9)
+  test_df = COMPLEX_AGENT_DATA.iloc[:0, :]
+  test_df_copy = test_df.copy(deep=True)
+  df = CLIENT.decide_from_contexts_df(tree, test_df)
+
+  assert_true(isinstance(df, pd.DataFrame))
+  assert_equal(len(df), 0)
+  assert_true(test_df.equals(test_df_copy))
 
 def setup_complex_agent_2_with_data():
   setup_complex_agent_2()
@@ -168,9 +182,11 @@ def test_decide_from_contexts_df_null_decisions():
     ],
     columns=["b", "tz"],
     index=pd.date_range("20130201", periods=2, freq="D").tz_localize("Europe/Paris"))
+  test_df_copy = test_df.copy(deep=True)
 
   df = CLIENT.decide_from_contexts_df(tree, test_df)
   assert_equal(len(df), 2)
+  assert_true(test_df.equals(test_df_copy))
   assert pd.isnull(df["a_predicted_value"][0])
   assert pd.notnull(df["error"][0])
 
@@ -193,9 +209,11 @@ def test_decide_from_contexts_df_with_array():
     ],
     columns=["b", "tz"],
     index=pd.date_range("20130201", periods=2, freq="D").tz_localize("Europe/Paris"))
+  test_df_copy = test_df.copy(deep=True)
 
   df = CLIENT.decide_from_contexts_df(tree, test_df)
   assert_equal(len(df), 2)
+  assert_true(test_df.equals(test_df_copy))
   assert pd.isnull(df["a_predicted_value"][0])
   assert pd.notnull(df["error"][0])
 
@@ -251,11 +269,13 @@ def test_datetime_decide_from_contexts_df():
     ],
     columns=["a"],
     index=pd.date_range("20130101 00:00:00", periods=3, freq="H").tz_localize("Asia/Shanghai"))
+  test_df_copy = test_df.copy(deep=True)
 
   df = CLIENT.decide_from_contexts_df(tree, test_df)
   assert_equal(len(df), 3)
   assert_equal(len(df.dtypes), 3)
   assert_equal(df["b_predicted_value"].tolist(), ["Pierre", "Paul", "Jacques"])
+  assert_true(test_df.equals(test_df_copy))
 
 @with_setup(setup_simple_agent_with_data, teardown)
 def test_tree_visualization():